add tools
