+++ public/index.html
@@ -17,10 +17,10 @@ initial-scale=1.0" />
       containerId: "assistant-agi", // id of the HTML tag
       apiUrl: "<ADD_YOUR_API_URL>", // server url or LangGraph cloud url
       apiKey: "<ADD_API_KEY>", // optional only for local development
-      configurable: { "param1": "value1" }, // LangGraph configurable parameters 
+      configurable: { "param1": "value1" }, // LangGraph configurable parameters
       headers: {
         'Authorization': 'Bearer <TOKEN>',
-      } // your server needed headers 
+      } // your server needed headers
     })
   </script>
 </body>
