+++ src/components/Assistant/types.ts
@@ -1,0 +1,3 @@
+import {ChatBotInput} from "../../types";
+
+export type AssistantParams = Omit<ChatBotInput, "containerId">;
