+++ src/components/tools/ActionTool/components/TransactionConfirmationPending/types.ts
@@ -1,0 +1,4 @@
+export type TransactionConfirmationPendingProps = {
+    onConfirm: () => void;
+    onReject: () => void;
+};
