+++ src/components/tools/ActionTool/types.ts
@@ -1,0 +1,5 @@
+export type ActionResult = {
+    approve?: boolean;
+    cancelled?: boolean;
+    error?: string;
+};
